requirements.txt
# Core scientific stack
TA-Lib>=0.6.5
akshare>=1.15.59
pandas>=2.2.0
numpy>=1.26.0

# Data processing
openpyxl>=3.1.0
pypdf>=3.0.0

# Web framework
requests>=2.31.0
Flask>=3.0.0
gunicorn
flask-cors>=4.0.0
supabase
python-dotenv
google-generativeai

# Plotting & Reporting (Restored for Render)
matplotlib>=3.8.0
mplfinance>=0.12.9b0
reportlab>=4.0.0
